fix: move `lottie-ios` to `peerDependencies`
`lottie-ios` is a project with native code and it has to be linked.

react-native autolinking can only work as long as project is listed in
the output of `react-native config` which considers `package.json` in
order to build list of dependencies which need to be linked.

currently having `lottie-ios` in `dependencies` does not make lots of
sense since in order to successfully install and use
`lottie-react-native` one has to install `lottie-ios` separately and
make sure it's listed in `package.json`.

in fact it may lead to situatoin when two copies of `lottie-ios` may be
installed in case if range of `lottie-ios` inside package.json of
`lottie-react-native` and inside actual package of the app are
conflicting.